Add a class instance test for ThankYouCard#setCanvasSize
We are using Enzyme's `.instance()` method to access the component's class
instance so we can test individual methods on the class. Note that as of
React 16, `instance()` returns null for functional components so only use
this strategy with class components.

For more info: https://airbnb.io/enzyme/docs/api/ReactWrapper/instance.html