Add some additional example unit tests.
1. #getTextStyleConfig - We're using the Jest matcher `.objectContaining(object)`
(https://jestjs.io/docs/en/expect#expectobjectcontainingobject)to test that
the shape of a returned object matches what we expect. This is a useful
test for testing the interface of your methods.

2. #createImageForCanvas - We're using the Jest matcher `toBeInstanceOf()`
to test the type of the return value

3. #drawBackgroundImage - We're using the Jest matcher `toHaveBeenCalledWith()`
to test that a function we expected to be called was. This might not always
be a great test as it's testing implementation details. But in some cases
can be useful.