Add test for clicking next and prev buttons
Following Kent C. Dodds philosophy (author of React Testing Library), we
want to test only the user visible results of events and actions.

To test what occurs when a user clicks the next button, we test that the
next thank you card is rendered.

In many unit tests, particulary Enzyme tests, you would see the test checking
the state of the component. This violates the philosophy of unit testing
since state is an internal implementation detail of the component.