Add tests for search functionality
Add tests to check that the correct thank you card is rendered when an
exact name match, close name match, and differently cased name matches.

Add a test to check that the error message is rendered when no match for
member name exists.

Again, we are testing the user visible results of events rather than internal
state of the component.